Separate out loop calculation functions, update smethod default param, updated README
